feat: complete Zhike tuning workflow + fix 144 compilation errors
- Fix all type mismatches: unify EnergyWh/EfficiencyWhKm as Float
- Add ZhikeSettingsValidator (13 validation rules: errors + warnings)
- Add ZhikeWorkflowModels (9-state write machine + protocol events)
- Add ZhikePostWriteVerifier (post-write read-back verification)
- Add AA210001 write mode handshake to BleManager/ZhikeProtocol
- Add ACK-driven protocol events (WriteModeReady/WriteSuccess/WriteFailure)
- Add zhikeWriteState StateFlow + executeZhikeWriteWorkflow in MainViewModel
- Update ZhikeSettingsScreen: write state UI, validation dialog, risk labels
- Add factory reset + bluetooth password change workflow
- Add parameter JSON import/export with file picker
- Update desktop issues document

+++ app/src/main/java/com/shawnrain/habe/data/RideSession.kt
@@ -2,10 +2,10 @@ package com.shawnrain.habe.data
 
 data class RideSession(
     val date: String = "",
-    val distanceKm: Double = 0.0,
+    val distanceKm: Float = 0.0f,
     val durationMin: Int = 0,
-    val maxSpeed: Float = 0f,
-    val avgSpeed: Float = 0f,
-    val totalWh: Float = 0f,
-    val avgEfficiency: Float = 0f
+    val maxSpeed: Float = 0.0f,
+    val avgSpeed: Float = 0.0f,
+    val totalWh: Float = 0.0f,
+    val avgEfficiency: Float = 0.0f
 )
